Introducing Training Portal
http://sync.cloudwick.com/training

SYNC-76 SYNC-77 SYNC-78 SYNC-79 #resolve

Allows users (in this case consultants/trainers/admins) to view training
content in a new and intutive way. Admins/Trainers will be able to
upload the training content, which automatically gets converted to
images and stored to Mongo to be served later by the presentation module
of the training portal. This provides security such that
trainers/trainees will not be able to save content like saveaspdf.

Presentation module of the training portal provides a simple interface
to present the content uploaded by the content manager. This module
provides the user to navigate around the content using arrow keys.